CLC-6200 Extract LDAP connectivity from CanvasCsv module
